Update directory structure.
