+++ pyproject.toml
@@ -6,31 +6,31 @@ build-backend = "setuptools.build_meta"
 name = "simple-mecab"
 description = "Use MeCab with simple code."
 readme = "README.md"
-license = {file = "LICENSE"}
-authors = [{ name = "Keisuke Magara", email = "Keisuke-Magara.dev@outlook.com"}]
+license = { file = "LICENSE" }
+authors = [
+    { name = "Keisuke Magara", email = "Keisuke-Magara.dev@outlook.com" },
+]
 keywords = ["mecab"]
 classifiers = [
     "Programming Language :: Python :: 3.10",
     "Intended Audience :: Science/Research",
     "License :: OSI Approved :: MIT License",
-    "Natural Language :: Japanese"
+    "Natural Language :: Japanese",
 ]
 requires-python = ">=3.7"
-dependencies = [
-    "mecab>=0.996"
-]
+dependencies = ["mecab>=0.996"]
 dynamic = ["version"]
 
 [project.urls]
 repository = "https://github.com/Keisuke-Magara/simple-mecab"
 
 [tool.setuptools]
-package-dir = {"" = "src"}
+package-dir = { "" = "src" }
 
 [tool.setuptools.dynamic]
-version = {attr = "setuptools_scm.get_version"}
+version = { attr = "simple_mecab.version" }
 
 [tool.setuptools_scm]
-write_to = "src/_version.py"
+write_to = "src/simple_mecab/_version.py"
 version_scheme = "guess-next-dev"
 local_scheme = "no-local-version"
