--- src/__init__.py
@@ -1,3 +1,0 @@
-from _version import version
-
-__version__ = version
