+++ src/simple_mecab/__init__.py
@@ -1,0 +1,6 @@
+from simple_mecab._version import version
+from simple_mecab.exceptions import *
+from simple_mecab.mecab_wrapper import *
+from simple_mecab.morpheme import *
+
+__version__ = version
