+++ src/simple_mecab/exceptions.py
@@ -1,0 +1,2 @@
+class InvalidArgumentsError(Exception):
+    pass
