+++ src/simple_mecab/mecab_wrapper.py
@@ -1,78 +1,10 @@
-"""形態素解析ライブラリMeCabを簡単に使えるようにするモジュールです。
-
-Requirements
-------------
-- Python 3.6 以上で動作します。
-- コンピュータに MeCab がインストールされている必要があります。
-- mecab ライブラリが必要です。
-"""
 import re
-from dataclasses import dataclass
-from typing import List, Literal, Optional
+from typing import List, Literal
 
 import MeCab
 
-
-@dataclass
-class Morph:
-    """MeCabの解析結果を要素別に格納するためのdataclassです。
-
-    Variables
-    ---------
-    word : str
-        形態素 （例：'食べ'）
-
-    pos0 : str | None
-        語の品詞 (part of speech)
-        （例：'動詞'）
-
-    pos1 : str | None
-        品詞細分類1
-        （例：'代名詞'）
-
-    pos2 : str | None
-        品詞細分類2
-        （例：'一般'）
-
-    pos3 : str | None
-        品詞細分類3
-        （例：'場所'）
-
-    conjugation_type : str | None
-        活用型
-        （例：'下一段-バ行'）
-
-    conjugation : str | None
-        活用形
-        （例：'未然形-一般'）
-
-    stem_form : str | None
-        原形
-        （例：'食べる'）
-
-    pronunciation : str | None
-        発音
-        （例：'タベ'）
-
-    unknown : str | None
-        正常に抽出できなかった場合はここに入ります。
-
-    **それぞれの要素に入る値は使用する辞書によって異なります。**
-    """
-    word: str
-    pos0: Optional[str]
-    pos1: Optional[str]
-    pos2: Optional[str]
-    pos3: Optional[str]
-    conjugation_type: Optional[str]
-    conjugation: Optional[str]
-    stem_form: Optional[str]
-    pronunciation: Optional[str]
-    unknown: Optional[str]
-
-
-class InvalidArgumentsError(Exception):
-    pass
+from simple_mecab.exceptions import InvalidArgumentsError
+from simple_mecab.morpheme import Morpheme
 
 
 class MeCabWrapper:
@@ -101,7 +33,7 @@ class MeCabWrapper:
 
     - `mecab-python3` ライブラリがインストールされている必要があります。
 
-    - 同ライブラリの中の Morph dataclass を使用して結果を格納します。
+    - 同ライブラリの中の Morpheme dataclass を使用して結果を格納します。
     """
 
     _none_pattern: List[str] = ['', ' ', '*']  # 該当なしのパターン
@@ -151,7 +83,7 @@ class MeCabWrapper:
                 f"{banned_args}\n"
                 "[ヒント] もし分かち書きをしたいのであれば、wakati_gaki関数を使用することができます。")
 
-    def parse(self, sentence: str) -> List[Morph]:
+    def parse(self, sentence: str) -> List[Morpheme]:
         """日本語の文字列をMeCabで解析します。
 
         Parameters
@@ -161,58 +93,57 @@ class MeCabWrapper:
 
         Returns
         -------
-        list[Morph]
-            形態素ごとにそれぞれ Morph クラスに情報が格納されています。
-            （アクセス例：`mecab_agent.parse()[0].word`）
-            詳細は Morph クラスの DocString を参照してください。
+        list[Morpheme]
+            形態素ごとにそれぞれ Morpheme クラスに情報が格納されています。
+            （アクセス例：`mecab_agent.parse()[0].token`）
+            詳細は Morpheme クラスの DocString を参照してください。
         """
-        result: list[Morph] = []
+        result: List[Morpheme] = []
         if sentence is not None:
             self.latest_input = sentence
         parsed_string = self.tagger.parse(self.latest_input)
-        words: list[str] = parsed_string.split('\n')
+        words: List[str] = parsed_string.split('\n')
         words.remove('EOS')
         words.remove('')
         for w in words:
             result.append(self._extract(w))
         return result
 
-    def _extract(self, parsed_word: str) -> Morph:
+    def _extract(self, parsed_word: str) -> Morpheme:
         if self.parse_type == 'ipadic':
             surface, others = parsed_word.split('\t')
             info = others.split(',')
-            ret = Morph(surface,
-                        info[0] if len(
-                            info) > 0 and info[0] not in self._none_pattern else None,
-                        info[1] if len(
-                            info) > 1 and info[1] not in self._none_pattern else None,
-                        info[2] if len(
-                            info) > 2 and info[2] not in self._none_pattern else None,
-                        info[3] if len(
-                            info) > 3 and info[3] not in self._none_pattern else None,
-                        info[4] if len(
-                            info) > 4 and info[4] not in self._none_pattern else None,
-                        info[5] if len(
-                            info) > 5 and info[5] not in self._none_pattern else None,
-                        info[6] if len(
-                            info) > 6 and info[6] not in self._none_pattern else None,
-                        info[7] if len(
-                            info) > 7 and info[7] not in self._none_pattern else None,
-                        None)
+            ret = Morpheme(surface,
+                           info[0] if len(
+                               info) > 0 and info[0] not in self._none_pattern else None,
+                           info[1] if len(
+                               info) > 1 and info[1] not in self._none_pattern else None,
+                           info[2] if len(
+                               info) > 2 and info[2] not in self._none_pattern else None,
+                           info[3] if len(
+                               info) > 3 and info[3] not in self._none_pattern else None,
+                           info[4] if len(
+                               info) > 4 and info[4] not in self._none_pattern else None,
+                           info[5] if len(
+                               info) > 5 and info[5] not in self._none_pattern else None,
+                           info[6] if len(
+                               info) > 6 and info[6] not in self._none_pattern else None,
+                           info[7] if len(
+                               info) > 7 and info[7] not in self._none_pattern else None,
+                           None)
             return ret
         elif self.parse_type == 'neologd':
             raise NotImplementedError(
-                "NEologd dictionary is not supported now. Please wait for version 2.0.0.")
+                "mecab-ipadic-NEologd辞書のパーサーは未実装です。")
         elif self.parse_type == 'unidic':
-            raise NotImplementedError(
-                "UniDic dictionary is not supported now. Please wait for version 2.0.0.")
+            raise NotImplementedError("UniDic辞書のパーサーは未実装です。")
         else:
-            word, other = parsed_word.split()
-            ret = Morph(word, None, None, None, None,
-                        None, None, None, None, other)
+            surface, feature = parsed_word.split()
+            ret = Morpheme(surface, None, None, None, None,
+                           None, None, None, None, feature)
             return ret
 
-    def wakati_gaki(self, sentence: str):
+    def wakati_gaki(self, sentence: str) -> str:
         """文を分かち書きして、リストに格納します。
 
         Parameters
@@ -225,7 +156,7 @@ class MeCabWrapper:
         list[str]
             分かち書きされた形態素のリスト
         """
-        wakati_list = []
+        wakati_list: List[str] = []
         for e in self.parse(sentence):
-            wakati_list.append(e.word)
-        return wakati_list
+            wakati_list.append(e.token)
+        return ' '.join(wakati_list)
